backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
psycopg2-binary
python-dotenv
Flask-Marshmallow
marshmallow-sqlalchemy